+++ packages/polymath-investor/src/app/sto/PurchaseForm.js
@@ -5,7 +5,11 @@ import React, { Component } from 'react';
 import { Field, reduxForm } from 'redux-form';
 import { Form, Button } from 'carbon-components-react';
 import { TextInput } from '@polymathnetwork/ui/deprecated';
-import { required, integer, float } from '@polymathnetwork/ui/validate';
+import {
+  required,
+  integer,
+  float,
+} from '@polymathnetwork/ui/deprecated/validate';
 import type { STODetails } from '@polymathnetwork/js';
 
 export const formName = 'purchase';
