+++ packages/polymath-issuer/src/pages/compliance/components/LockupDatesFields.js
@@ -3,7 +3,10 @@
 import React, { Component } from 'react';
 import { Field } from 'redux-form';
 import { CheckboxInput, DatePickerInput } from '@polymathnetwork/ui/deprecated';
-import { required, tomorrowOrLater } from '@polymathnetwork/ui/validate';
+import {
+  required,
+  tomorrowOrLater,
+} from '@polymathnetwork/ui/deprecated/validate';
 
 type Props = {|
   isSalePermanent: boolean,
