+++ packages/polymath-issuer/src/pages/providers/ApplyForm.js
@@ -4,7 +4,11 @@ import React, { Component } from 'react';
 import { Field, reduxForm } from 'redux-form';
 import { Form, Tooltip, Button } from 'carbon-components-react';
 import { TextInput, TextAreaInput } from '@polymathnetwork/ui/deprecated';
-import { required, maxLength, url } from '@polymathnetwork/ui/validate';
+import {
+  required,
+  maxLength,
+  url,
+} from '@polymathnetwork/ui/deprecated/validate';
 
 export const formName = 'providers_apply';
