+++ packages/polymath-ui/src/components/SignUpPage/SignUpForm.js
@@ -5,7 +5,7 @@ import { Form, Button, Link } from 'carbon-components-react';
 import { Field, reduxForm } from 'redux-form';
 
 import { CheckboxInput, TextInput } from '../../deprecated/components/inputs';
-import { required, maxLength, email } from '../../validate';
+import { required, maxLength, email } from '../../deprecated/validate';
 
 export const formName = 'signup';
